Fix issue with output token decimals
The output token amount from the swap was using the input token's decimals. Would work of if they matches but fails for tokens like USDC to SOL swaps. This fixes it to use the correct decimal number for the correct display of the swap amount on screen to the user.