Update index.ts
Added in the new transaction retry logic (v1) that simply pauses, fetches a new route list and tries again (top route) up to the max set in the .env variable. Mixing this logic with using existing route #1 and #2 may be a good middle ground. Also added some styling and a few more bonuses here to make it better. Also limited the Jupiter top result to 1 to add some efficiency with the list.